feat: add configurable temperature and humidity offsets in DHT sensor and improve float precision handling

+++ esp8266-binary-clock/MyMQTT.h
@@ -55,7 +55,7 @@ class MyMQTT : public WithTicker {
         void onMqttDisconnectTick_();
         void disconnectWifi_();
         void publish_();
-        void publishValue_(const char* subtopic, int value);
+        void publishValue_(const char* subtopic, float value);
 };
 
 #endif // MYMQTT_H
